+++ core/dbt/adapters/exceptions/__init__.py
@@ -1,2 +1,4 @@
 from dbt.adapters.exceptions.compilation import *  # noqa
 from dbt.adapters.exceptions.alias import *  # noqa
+from dbt.adapters.exceptions.database import *  # noqa
+from dbt.adapters.exceptions.connection import *  # noqa
