+++ core/dbt/config/profile.py
@@ -16,7 +16,7 @@ from dbt.exceptions import (
     DbtRuntimeError,
     ProfileConfigError,
 )
-from ..common.exceptions import DbtValidationError
+from dbt.common.exceptions import DbtValidationError
 from dbt.common.events.types import MissingProfileTarget
 from dbt.common.events.functions import fire_event
 from dbt.utils import coerce_dict_str
