+++ core/dbt/task/run.py
@@ -25,8 +25,8 @@ from dbt.exceptions import (
     DbtInternalError,
     DbtRuntimeError,
 )
-from ..common.exceptions import DbtValidationError
-from ..adapters.exceptions import MissingMaterializationError
+from dbt.common.exceptions import DbtValidationError
+from dbt.adapters.exceptions import MissingMaterializationError
 from dbt.common.events.functions import fire_event, get_invocation_id
 from dbt.common.events.types import (
     DatabaseErrorRunningHook,
