+++ plugins/postgres/dbt/adapters/postgres/impl.py
@@ -12,7 +12,7 @@ from dbt.adapters.postgres import PostgresRelation
 from dbt.common.dataclass_schema import dbtClassMixin, ValidationError
 from dbt.common.exceptions import DbtRuntimeError
 from dbt.contracts.graph.nodes import ConstraintType
-from dbt.exceptions import (
+from dbt.adapters.exceptions import (
     CrossDbReferenceProhibitedError,
     IndexConfigNotDictError,
     IndexConfigError,
