+++ tests/functional/partial_parsing/test_pp_vars.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 from dbt.constants import SECRET_ENV_PREFIX
 from dbt.exceptions import ParsingError
-from dbt.adapters.exceptions.connection import FailedToConnectError
+from dbt.adapters.exceptions import FailedToConnectError
 from dbt.tests.util import get_manifest, run_dbt, run_dbt_and_capture, write_file
 
 from tests.functional.partial_parsing.fixtures import (
